feat: Add functionality to perform house cleaning plan
This commit introduces the ability to execute a house cleaning plan using Google Cloud services. The new command "house_cleaning" has been added to the main script. It allows the user to run a house cleaning plan by accessing the Google Drive metadata and interacting with a Google Sheets document.